Extract garray.c from misc2.c
Start to split misc2.c in many other files (see #209).
