+++ src/charset.c
@@ -14,6 +14,7 @@
 #include "memline.h"
 #include "misc1.h"
 #include "misc2.h"
+#include "garray.h"
 #include "move.h"
 #include "os_unix.h"
