+++ src/ex_docmd.c
@@ -37,6 +37,7 @@
 #include "message.h"
 #include "misc1.h"
 #include "misc2.h"
+#include "garray.h"
 #include "move.h"
 #include "normal.h"
 #include "ops.h"
