+++ src/ex_getln.c
@@ -32,6 +32,7 @@
 #include "message.h"
 #include "misc1.h"
 #include "misc2.h"
+#include "garray.h"
 #include "move.h"
 #include "ops.h"
 #include "option.h"
