+++ src/hardcopy.c
@@ -25,6 +25,7 @@
 #include "message.h"
 #include "misc1.h"
 #include "misc2.h"
+#include "garray.h"
 #include "option.h"
 #include "screen.h"
 #include "syntax.h"
