+++ src/menu.c
@@ -21,6 +21,7 @@
 #include "message.h"
 #include "misc1.h"
 #include "misc2.h"
+#include "garray.h"
 #include "term.h"
