+++ src/misc1.c
@@ -30,6 +30,7 @@
 #include "memline.h"
 #include "message.h"
 #include "misc2.h"
+#include "garray.h"
 #include "move.h"
 #include "option.h"
 #include "os_unix.h"
