+++ src/os_unix.c
@@ -41,6 +41,7 @@
 #include "message.h"
 #include "misc1.h"
 #include "misc2.h"
+#include "garray.h"
 #include "screen.h"
 #include "syntax.h"
 #include "term.h"
