+++ src/spell.c
@@ -314,6 +314,7 @@
 #include "message.h"
 #include "misc1.h"
 #include "misc2.h"
+#include "garray.h"
 #include "normal.h"
 #include "option.h"
 #include "os_unix.h"
