+++ src/term.c
@@ -34,6 +34,7 @@
 #include "getchar.h"
 #include "message.h"
 #include "misc2.h"
+#include "garray.h"
 #include "move.h"
 #include "normal.h"
 #include "option.h"
