+++ src/ui.c
@@ -25,6 +25,7 @@
 #include "mbyte.h"
 #include "misc1.h"
 #include "misc2.h"
+#include "garray.h"
 #include "move.h"
 #include "normal.h"
 #include "option.h"
