+++ src/undo.c
@@ -92,6 +92,7 @@
 #include "message.h"
 #include "misc1.h"
 #include "misc2.h"
+#include "garray.h"
 #include "option.h"
 #include "os_unix.h"
 #include "quickfix.h"
